refactor: consolidating the `internal` package
